Finish cp, mv, rm, and clean up other subcmds
Next step is to change encryption to use PASSWORD_STORE_SIGNING_KEY
(Vec<String>) to detach-sign every encrypted file -- if specified not
empty. I also need to not /encrypt/ with those keys -- just with the
one(s) specified in the .gpg-id file (which needs to be verified for
integrity before being read) and/or PASSWORD_STORE_KEY (Vec<String>).

Could probably reuse the `recrypt_store` function to read all the
encryption keys into a Vec<String>.