Second cleaning pass
I just realized the pun in this and the prior commit titles. Haha.

* I decided to move the PassrsError enum (name subject to change) back
to its own module, but instead re-export the enum from main.rs. I forgot
this was a thing, which is why I moved it to main.rs originally: less
keystrokes to import my errors.

* I changed PASSWORD_STORE_DIR to be a PathBuf so I wouldn't have to do
that janky if-ends-with-slash test.

* I added two constants, PASSWORD_STORE_LEN and PASSWORD_STORE_STRING. I
don't know if this makes a difference and saves on allocations, but it
looks nicer than `PASSWORD_STORE_DIR.display().to_string().len()`.

* `gpg_id(s)` is no longer required for the init subcommand, in
preparation of making an empty ID wipe out the store -- see
/usr/bin/pass:317. I'll add a prompt, though, because the lack of user
interaction on such a destructive operation doesn't seem like a good
idea (especially when I'm mostly reimplementing somebody else's
program and am bound to have made mistakes).

* Moved all clipboard logic to clipboard.rs -- now it actually handles
clipping things.

* Lots of other little things, such as moving all clipboard logic to
clipboard.rs.

I still need to come to a decision on whether or not I will support
git-less initialization. Do people even use bare password stores? To be
continued.