Third cleaning pass
This is shaping up nicely. Hopefully I'll be able to get this out before
classes start. I'd still like to set up a test suite akin to what pass
has, as well as allow the autogenerated completions to support the
completion of filenames from the store. Before that, however, I need to
work on documenting everything, fleshing out the README, and start using
this in place of {go,}pass to ensure it is actually usable day-to-day.

* I decided against allowing the option of wiping the store and have
thus reinstated the `gpg_id` requirement of the `init` subcmd. My
rationale behind this is that 1) pass requires an empty `gpg_id(s)` arg
(not missing, but empty -- ""); and 2) if you were to deinit your store
right this moment, it would delete the root .gpg-id file, throw an error
that you must initialize your password store before using it, and print
out help. Basically, it removes one single file, commits that, and spews
a whole bunch of stuff to stdout. I'd be more inclined to support
reinitialization (with an empty `gpg_id`) than this. But for now,
neither of those are supported.

* Show similar output to when `git commit` is used, showing the first 7
digits of the commit hash, the commit message, followed by a summarized,
shortstat'd diff. Similarity is not yet exposed in `git2-rs`, so it will
always display 100% similarity for now. The rename detection isn't as
sophisticated as `git`, so sometimes it might appear as "deleted file,
added file" when using the `mv` subcommand (that, or I've messed
something up).

* If the user has `commit.gpgsign` set to `true` in their gpg config,
sign commits. Otherwise, don't.

* Drop `rand` as a dependency and use `ring::rand` for random
generation.

* Add `libc` as a dependency so I can get the user's uid and compare
against the specified path passed to `util::set_recursive_permissions`
in order to check if the user owns a path before setting
permissions (would otherwise error if the user didn't own the path).

* Add `psutil` as a dependency so I can kill any previously-running
instances of pass-rs when utilizing the clipboard (to prevent races to
clear the clipboard -- e.g. you clip something, then clip something
else, the first instance of `unclip` would still clear the clipboard
after the timeout, thus clearing the newly clipped contents).

* Fix PASSWORD_STORE_SIGNING_KEY implementation: it is only used to sign
the `.gpg-id` file, not individual entries to the store.

* Allow the generation of a first-time OTP code via the `--generate` (or
`-g`) flag when using `otp::append` or `otp::insert`, to allow for
easier verification of the code. Additionally, display the time left on
that specific code with a fancy little "progress" bar using `format!`
magic (not really).

* Create a `Flags` struct for functions that take more than one boolean
flag because clippy was complaining that 8 is too many args for
`otp::insert`.

* Remove leftover empty dirs when using the `mv` subcmd if there are no
other files in that dir.

* Some refactoring of the `init` subcmd, moving `recrypt_{file,dir}`
into the util module.

* Add licenses for my use of code from `tui-rs`, `treeline`, `copy_dir`,
`grep-cli`, and `terminal_qrcode`.