Fourth cleaning pass
The last few steps I need to deal with are: completing the README,
finishing the completions for `bash` and `zsh`, and fixing whatever bugs
crop up as I notice them.

* `structopt` finally cut a new release, so no more git dependencies!
* Started working on the README to somewhat resemble that of a project
and not vaporware.
* Started working on completions -- the `fish` ones are mostly done. The
`bash` and `zsh` ones are autogenerated by `clap`. They probably work,
but they won't complete secret keys, or directories and entries from the
store.
* Did some documentation. It looks alright. The wording is a bit awkward
in places, but it is my first time actually writing documentation, so
cut me some slack.
* Gave docstrings to all subcommands/flags/arguments so everything
should be explained somewhat decently.
* Removed the `complete` subcommand in favor of manually expanding on
the autogenerated completions. More of a maintenance burden, but it's
worth it if it allows completing entries in the store and similar.
* Also moved arguments around -- `secret_name` will always be first, and
any flags will always be last. Everything else goes in-between.
* To avoid adding unrelated files to git, the `util::commit` function
now takes an `Option<&[&Path]>` -- if `None`, just update whatever has
changed (and therefore not adding unrelated files); if `Some(&[..])`,
add those paths to the index. This has the side effect of requiring all
`None`s to be annotated as `None::<&[PathBuf]>` which looks ugly but
whatever.
* Renamed `util::find_target_single` to `util::find_matches` because I
removed `util::find_target_multi` a long time ago (which was the reason
for the name).
* Replaced all `unwrap`s with `expect`s -- if they happen even though
the shouldn't, this will make it easier to track down.
* Made all `collect::<Vec<_>>`s more explicit by giving the variables
explicit types.
* `mv` can now move a file into a directory -- e.g. `passrs mv test a/`
would result in `a/test.gpg` being created, whereas before it would
create `a/.gpg`.
* Also, don't recrypt `mv`'d files unless the closest `.gpg-id` files
are different (meaning they have different encryption keys).
* Made `util::copy` a helper that calls out to the real
`util::copy_impl` to eliminate the `None` at every call.
* Implemented `Display` for the `Cell` struct to make printing it look
nicer (to me, at least).
* A little bit of cleaning up `tree` and some renaming -- now everything
should be in alphabetical order thanks to some nifty derives. However, I
had to go back to a `Vec<Tree>` in order to actually sort it (`&mut
[Tree]` wouldn't work because of the `Display` impl only using `&self`).
* Change up `util::canonicalize_path` yet again; short-ciruit if a file
exists at `path + ".gpg"`, otherwise continue on down the conditionals.
* Create a test script -- the tests shouldn't fail if they are run on
the supplied test store.