util: fix matching exact entries
If you had one entry in your store, `a.gpg`, `passrs show a` would not
work because the filename includes the extension, while the query did
not. `passrs show a.gpg` would work as expected. Now, `passrs show a`
and `passrs show a.gpg` will both work. This might introduce some more
bugs, but those can be dealt with at a later date.