cli: don't require tty for edit
This allows users to run

    EDITOR=cat passrs edit some/secret/file | grep something

in order to grep for a string in a secret file (for example).

I went this route rather than modify `passrs show` because `show` is
special-cased to only show the first line when being piped (to better
support browser addons like passff), and I feel it's worth it to keep
that.